chatbot/travel_api/requirements.txt
google-cloud-bigquery
functions-framework==3.*
firebase-functions==0.1.*
requests==2.*
google-auth==2.*